requirements.txt
fastapi
uvicorn
motor==3.5.3
pydantic==1.10.8
python-multipart
bcrypt
pymongo[srv]==4.6.1
jinja2
altair==4.2.0
python-dotenv
python-jose